Reorganize login flow: Authentications onl starts with queryString code and no token.  Otherwise redirect to /profile Closes 56
