+++ src/components/Login/components/GithubLoginModal.jsx
@@ -1,0 +1,37 @@
+import React from "react"
+
+// -- COMPONENTS -- //
+// TODO: refactor styles
+const GithubLoginModal = ({
+  redirect = "/profile",
+  clientID = "e015fd9cc874fa5a34bf"
+}) => {
+  const githubAuthURL = `https://github.com/login/oauth/authorize?client_id=${clientID}`;
+  // TODO: process.env for API and base client URL flip on dev / prod (default prod)
+  // const redirectBase = `${procescs.env.CLIENT_URL}/login`;
+  const redirectBase = "https://localhost:3000/login"
+  // const { redirect } = qs.parse(queryString);
+  return (
+    <div className="login-box" onClick={e => e.stopPropagation()} >
+      <div className="login-title">Github Authentication</div>
+      <a
+        className="login-link"
+        href={
+          redirect
+            ? githubAuthURL
+            : `${githubAuthURL}&redirect_uri=${redirect}`
+        }>
+        <button className="github-auth">
+          <img
+            alt="github-icon"
+            className="github-icon"
+            src={"https://i.imgur.com/UBZgVgQ.png"}
+          />
+          Log in with Github
+      </button>
+      </a>
+    </div >
+  )
+};
+
+export default GithubLoginModal
